cfo_dashboard/requirements.txt
Django==4.2.7
pandas==2.2.3
pdfplumber==0.11.4
numpy==2.1.2
plotly==5.24.1
yfinance==0.2.50
requests==2.32.3
Pillow==11.0.0
python-decouple==3.8
celery==5.4.0
redis==5.1.1

# scikit-learn==1.3.0  # Add this for enhanced forecasting
scipy==1.11.3 